move fileloader to its own file
