+++ go.mod
@@ -18,7 +18,6 @@ require (
 	github.com/mattn/go-colorable v0.1.7 // indirect
 	github.com/mgutz/ansi v0.0.0-20200706080929-d51e80ef957d // indirect
 	github.com/niemeyer/pretty v0.0.0-20200227124842-a10e7caefd8e // indirect
-	github.com/pkg/errors v0.9.1 // indirect
 	github.com/satori/go.uuid v1.2.0
 	github.com/sirupsen/logrus v1.4.3-0.20191026113918-67a7fdcf741f
 	github.com/x-cray/logrus-prefixed-formatter v0.5.2
